Define "fetch" feature
Define and implement a "fetch" feature to better separate the logic
related to the fetching of the advisory database from the matching of
vulnerabilities against that database. This feature is enabled by
default to conserve the current behaviour.

This change defines a new Repository trait which only exposes the
required functions (as used by the Database struct). The Git-specific
implementation is moved to a GitRepository struct, with the logic
unchanged.

Drop the Database.fetch method. It is not used by cargo-audit and
increases the dependency on Repository's internals.